cloud-resume-api/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.16.1
Flask==3.0.0